Data.DNN.Trainer

Description

This module implements a monad for training a DNN using back-propagation or contrastive divergence.

Documentation

type Trainer m a = ExceptT a (StateT DNNS m) a

training monad type

finish :: (Monad m, MonadError a m, MonadState DNNS m) => a -> m ()

terminate the training script

finish_ :: (Monad m, MonadError () m, MonadState DNNS m) => m ()

terminate the training script with ()

run :: Monad m => DNN -> Trainer m a -> m (a, DNN)

Run the script over the DNN.

backward :: (Monad m, MonadError a m, MonadState DNNS m) => Matrix U B H -> m (Matrix U B I)

Run RBM algorithm backward

feedForward :: (Monad m, MonadError a m, MonadState DNNS m) => Matrix U B I -> m (Matrix U B H)

Run feedForward MLP algorithm over the entire DNN.

backProp :: (Monad m, MonadError a m, MonadState DNNS m) => Matrix U B I -> Matrix U B H -> m ()

Run Back Propagation training over the entire DNN.

contraDiv :: (Monad m, MonadError a m, MonadState DNNS m) => Matrix U B I -> m ()

Run Constrastive Divergance on the last layer in the DNN |This will first run the input through the layers to generate |the probability vector as the input for the last layer.

forwardErr :: (Monad m, MonadError a m, MonadState DNNS m) => Matrix U B I -> Matrix U B H -> m Double

Run feedForward MLP algorithm over the entire DNN.

reconErr :: (Monad m, MonadError a m, MonadState DNNS m) => Matrix U B I -> m Double

Compute the input reconstruction error with the current RBM in the state.

reconstruct :: (Monad m, MonadError a m, MonadState DNNS m) => Matrix U B I -> m (Matrix U B I)

Reconstruct the input with the current RBM

getCount :: (Monad m, MonadError a m, MonadState DNNS m) => m Int

Return how many times we have executed contraDiv or backProp

incCount :: (Monad m, MonadError a m, MonadState DNNS m) => m Int

Increment the count and the return the previous value.

putDNN :: (Monad m, MonadError a m, MonadState DNNS m) => [Matrix U I H] -> m ()

Set the DNN

getDNN :: (Monad m, MonadError a m, MonadState DNNS m) => m [Matrix U I H]

Return the DNN

popLastLayer :: (Monad m, MonadError a m, MonadState DNNS m) => m (Matrix U I H)

Pop the last layer of the DNN.

pushLastLayer :: (Monad m, MonadError a m, MonadState DNNS m) => Matrix U I H -> m ()

Push the updated layer as the last in the DNN.

nextSeed :: (Monad m, MonadError a m, MonadState DNNS m) => m Int

Get the next random seed.

setLearnRate :: (Monad m, MonadError a m, MonadState DNNS m) => Double -> m ()

Set the learning rate used in backProp and contraDiv.

getLearnRate :: (Monad m, MonadError a m, MonadState DNNS m) => m Double

Get the current learning rate.