Data.Matrix

Description

This module implements some matrix operations using the Repa package that track the symbolic shape of the matrix.

Documentation

data Matrix d a b

wraps the Repa Array types so we can typecheck the results of | the matrix operations

Constructors

Source d Double => Matrix (Array d DIM2 Double)

class MatrixOps a b where

Class implementing the typesafe Matrix apis

Minimal complete definition

Nothing

Methods

mmult :: Monad m => Matrix U a b -> Matrix U b c -> m (Matrix U a c)

mmultT :: Monad m => Matrix U a b -> Matrix U c b -> m (Matrix U a c)

d2u :: Monad m => Matrix D a b -> m (Matrix U a b)

(*^) :: Matrix c a b -> Matrix d a b -> Matrix D a b

(+^) :: Matrix c a b -> Matrix d a b -> Matrix D a b

(-^) :: Matrix c a b -> Matrix d a b -> Matrix D a b

map :: (Double -> Double) -> Matrix c a b -> Matrix D a b

cast1 :: Matrix c a b -> Matrix c d b

cast2 :: Matrix c a b -> Matrix c a d

transpose :: Monad m => Matrix U a b -> m (Matrix U b a)

sum :: Monad m => Matrix c a b -> m Double

mse :: Monad m => Matrix c a b -> m Double

elems :: Matrix c a b -> Int

row :: Matrix c a b -> Int

col :: Matrix c a b -> Int

shape :: Matrix c a b -> (Int, Int)

randomish :: (Int, Int) -> (Double, Double) -> Int -> Matrix U a b

splitRows :: Int -> Matrix c a b -> [Matrix D a b]

zipWith :: (Double -> Double -> Double) -> Matrix c a b -> Matrix c a b -> Matrix D a b

fromList :: (Int, Int) -> [Double] -> Matrix U a b

traverse :: (Double -> Int -> Int -> Double) -> Matrix c a b -> Matrix D a b

toList :: Matrix U a b -> [Double]

fold :: Monad m => (Double -> Double -> Double) -> Double -> Matrix U a b -> m Double

toUnboxed :: Matrix U a b -> Vector Double

data U :: *

Unboxed arrays are represented as unboxed vectors.

The implementation uses Data.Vector.Unboxed which is based on type families and picks an efficient, specialised representation for every element type. In particular, unboxed vectors of pairs are represented as pairs of unboxed vectors. This is the most efficient representation for numerical data.

data D :: *

Delayed arrays are represented as functions from the index to element value.

Every time you index into a delayed array the element at that position is recomputed.

data B

num inputs in batch

data I

num input nodes

data H

num hidden nodes