Data.RBM

Description

This module implements the Contrastive Diverence training algorithm for Restricted Boltzman Machines.

Documentation

new :: Int -> Int -> Int -> RBM

Create an rbm with some randomized weights

type RBM = Matrix U I H

weight matrix, numInputs x numHidden

where the bias node is the first node

contraDiv :: Monad m => Double -> Matrix U I H -> Int -> Matrix U B I -> m (Matrix U I H)

Run Contrastive Divergance learning. Return the updated RBM

energy :: Monad m => RBM -> Matrix U B I -> m Double

Compute the energy of the RBM with the batch of input.

hiddenPs :: Monad m => RBM -> Matrix U B I -> m (Matrix U B H)

Given a biased input generate probabilities of the hidden layer

incuding the biased probability.

inputPs :: Monad m => RBM -> Matrix U B H -> m (Matrix U I B)

Given a batch biased hidden sample generate probabilities of the input layer

incuding the biased probability.

sample :: Monad m => Int -> Matrix U a b -> m (Matrix U a b)

sample the matrix of probabilities

reconstruct :: Monad m => Matrix U B I -> [RBM] -> m (Matrix U B I)

Reconstruct the input by folding it forward over the stack of RBMs then backwards.

forward :: Monad m => Matrix U B I -> RBM -> m (Matrix U B I)

Run the RBM forward

backward :: Monad m => Matrix U B H -> RBM -> m (Matrix U B H)

Run the RBM backward