Examples.Mnist

Description

This module implements an example traning the MNIST data set. Using the dataset parsing code from: https://github.com/mhwombat/backprop-example/blob/master/Mnist.hs

Documentation

generateTrainBatches :: IO ()

Generate a batch of images with bias values for training.

generateTrainLabels :: IO ()

Generate a batch of of matching labels with bias values.

generateTestBatches :: IO ()

Generate a the test images with bias values sorted by label.

mnist :: IO ()

Train a network to recognize digits from the MNIST dataset. Start by training each layer as an RBM with the Contrastive Divergence algorithm. Then finish up the traning with Back-propagation. Animate the weights as the traning progresses.